+++ lib/generators/godmin/resource/resource_generator.rb
@@ -4,9 +4,13 @@ class Godmin::ResourceGenerator < Godmin::Generators::NamedBase
   argument :attributes, type: :array, default: [], banner: "attribute attribute"
 
   def add_route
-    inject_into_file "config/routes.rb", after: "godmin do\n" do
-      <<-END.strip_heredoc.indent(4)
-        resources :#{file_name.pluralize}
+    route "resources :#{file_name.pluralize}"
+  end
+
+  def add_navigation
+    append_to_file File.join("app/views", namespaced_path, "shared/_navigation.html.erb") do
+      <<-END.strip_heredoc
+        <%= navbar_item #{class_name} %>
       END
     end
   end