+++ lib/godmin.rb
@@ -8,7 +8,6 @@ require "godmin/authentication"
 require "godmin/authorization"
 require "godmin/engine"
 require "godmin/paginator"
-require "godmin/rails"
 require "godmin/resolver"
 require "godmin/resources/resource_controller"
 require "godmin/resources/resource_service"
@@ -17,9 +16,6 @@ require "godmin/version"
 module Godmin
   mattr_accessor :namespace
   self.namespace = nil
-
-  mattr_accessor :resources
-  self.resources = []
 
   def self.configure
     yield self